@@ -8,7 +8,7 @@ import 'package:flutter_modular/flutter_modular.dart';
 class OnBoardingStartPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    final localizations = AppLocalizations.of(context);
+    final localizations = AppLocalizations.of(context)!;
 
     return BasePage(
       scrollView: false,
