@@ -14,7 +14,7 @@ import 'intl/messages_all.dart';
 class S {
   S();
 
-  static S current;
+  static S? current;
 
   static const AppLocalizationDelegate delegate = AppLocalizationDelegate();
 
@@ -27,11 +27,11 @@ class S {
       Intl.defaultLocale = localeName;
       S.current = S();
 
-      return S.current;
+      return S.current!;
     });
   }
 
-  static S of(BuildContext context) {
+  static S? of(BuildContext context) {
     return Localizations.of<S>(context, S);
   }
 }
@@ -57,11 +57,9 @@ class AppLocalizationDelegate extends LocalizationsDelegate<S> {
   bool shouldReload(AppLocalizationDelegate old) => false;
 
   bool _isSupported(Locale locale) {
-    if (locale != null) {
-      for (var supportedLocale in supportedLocales) {
-        if (supportedLocale.languageCode == locale.languageCode) {
-          return true;
-        }
+    for (var supportedLocale in supportedLocales) {
+      if (supportedLocale.languageCode == locale.languageCode) {
+        return true;
       }
     }
     return false;
